Fixed floating point error

+++ result_checker.py
@@ -3,7 +3,7 @@ def resultCheck(user_answer, correct_answer):
     wrong_result = 0
     for i in range(len(user_answer)):
         for j in range(len(user_answer)):
-            if int(user_answer[i][j]) == correct_answer[i][j]:
+            if float(user_answer[i][j]) == correct_answer[i][j]:
                 correct_result += 1
             else:
                 wrong_result += 1
